# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API order management and tracking.",
        "description": "This API allows you to manage and track orders",
        "version": "1.0.0"
    },
    "paths": {
        "/api": {},
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get paginated orders",
                "description": "Returns a paginated list of orders",
                "operationId": "getOrders",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8001/api/orders?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8001/api/orders?page=1"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost:8001/api/orders"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Add a new Order",
                "operationId": "e1c8c4dbcb1f60187195df4791c9e8ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "total_amount",
                                    "items"
                                ],
                                "properties": {
                                    "total_amount": {
                                        "description": "Total amount of the order",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "items": {
                                        "description": "List of order items",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/OrderItem"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order was added successfuly",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Order": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{orderNumber}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by order number",
                "operationId": "2e24eb608bec8753fc59abd891344a6f",
                "parameters": [
                    {
                        "name": "orderNumber",
                        "in": "path",
                        "description": "Order Number",
                        "required": true,
                        "schema": {
                            "type": "srting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Order": {
                "required": [
                    "id",
                    "total_amount",
                    "status",
                    "items"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "order_number": {
                        "type": "string"
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "cancelled"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "order_items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "properties": {
                    "order_id": {
                        "type": "integer"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Operations related to orders"
        }
    ]
}